Move creation code in own extension file.
